Change fs mock interface
Allowing to mock once does not work well for procedures as scope delimiter.
File may be accessed more than once during the invocation of the tested
procedure, and this fact is an implementation detail subject to change.
Instead delimit the scope of the test by resetting the mocks before every test.